ES6 modules conversion
