+++ src/gl-matrix/common.js
@@ -19,52 +19,49 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE. */
 
 /**
- * @class Common utilities
- * @name glMatrix
+ * Common utilities
+ * @module glMatrix
  */
-var glMatrix = {};
 
 // Configuration Constants
-glMatrix.EPSILON = 0.000001;
-glMatrix.ARRAY_TYPE = (typeof Float32Array !== 'undefined') ? Float32Array : Array;
-glMatrix.RANDOM = Math.random;
-glMatrix.ENABLE_SIMD = false;
+export const EPSILON = 0.000001;
+export let ARRAY_TYPE = (typeof Float32Array !== 'undefined') ? Float32Array : Array;
+export const RANDOM = Math.random;
+export const ENABLE_SIMD = false;
 
 // Capability detection
-glMatrix.SIMD_AVAILABLE = (glMatrix.ARRAY_TYPE === this.Float32Array) && ('SIMD' in this);
-glMatrix.USE_SIMD = glMatrix.ENABLE_SIMD && glMatrix.SIMD_AVAILABLE;
+export const SIMD_AVAILABLE = (ARRAY_TYPE === Float32Array) && (typeof SIMD !== 'undefined');
+export const USE_SIMD = ENABLE_SIMD && SIMD_AVAILABLE;
 
 /**
  * Sets the type of array used when creating new vectors and matrices
  *
  * @param {Type} type Array type, such as Float32Array or Array
  */
-glMatrix.setMatrixArrayType = function(type) {
-    glMatrix.ARRAY_TYPE = type;
+export function setMatrixArrayType(type) {
+    ARRAY_TYPE = type;
 }
 
-var degree = Math.PI / 180;
+const degree = Math.PI / 180;
 
 /**
 * Convert Degree To Radian
 *
 * @param {Number} a Angle in Degrees
 */
-glMatrix.toRadian = function(a){
+export function toRadian(a){
      return a * degree;
 }
 
 /**
  * Tests whether or not the arguments have approximately the same value, within an absolute
- * or relative tolerance of glMatrix.EPSILON (an absolute tolerance is used for values less 
+ * or relative tolerance of glMatrix.EPSILON (an absolute tolerance is used for values less
  * than or equal to 1.0, and a relative tolerance is used for larger values)
- * 
+ *
  * @param {Number} a The first number to test.
  * @param {Number} b The second number to test.
  * @returns {Boolean} True if the numbers are approximately equal, false otherwise.
  */
-glMatrix.equals = function(a, b) {
-	return Math.abs(a - b) <= glMatrix.EPSILON*Math.max(1.0, Math.abs(a), Math.abs(b));
+export function equals(a, b) {
+	return Math.abs(a - b) <= EPSILON*Math.max(1.0, Math.abs(a), Math.abs(b));
 }
-
-module.exports = glMatrix;
