+++ src/gl-matrix/vec2.js
@@ -18,21 +18,20 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE. */
 
-var glMatrix = require("./common.js");
+import { ARRAY_TYPE, EPSILON, RANDOM } from "./common.js";
 
 /**
- * @class 2 Dimensional Vector
- * @name vec2
+ * 2 Dimensional Vector
+ * @module vec2
  */
-var vec2 = {};
 
 /**
  * Creates a new, empty vec2
  *
  * @returns {vec2} a new 2D vector
  */
-vec2.create = function() {
-    var out = new glMatrix.ARRAY_TYPE(2);
+export function create() {
+    var out = new ARRAY_TYPE(2);
     out[0] = 0;
     out[1] = 0;
     return out;
@@ -44,8 +43,8 @@ vec2.create = function() {
  * @param {vec2} a vector to clone
  * @returns {vec2} a new 2D vector
  */
-vec2.clone = function(a) {
-    var out = new glMatrix.ARRAY_TYPE(2);
+export function clone(a) {
+    var out = new ARRAY_TYPE(2);
     out[0] = a[0];
     out[1] = a[1];
     return out;
@@ -58,8 +57,8 @@ vec2.clone = function(a) {
  * @param {Number} y Y component
  * @returns {vec2} a new 2D vector
  */
-vec2.fromValues = function(x, y) {
-    var out = new glMatrix.ARRAY_TYPE(2);
+export function fromValues(x, y) {
+    var out = new ARRAY_TYPE(2);
     out[0] = x;
     out[1] = y;
     return out;
@@ -72,7 +71,7 @@ vec2.fromValues = function(x, y) {
  * @param {vec2} a the source vector
  * @returns {vec2} out
  */
-vec2.copy = function(out, a) {
+export function copy(out, a) {
     out[0] = a[0];
     out[1] = a[1];
     return out;
@@ -86,7 +85,7 @@ vec2.copy = function(out, a) {
  * @param {Number} y Y component
  * @returns {vec2} out
  */
-vec2.set = function(out, x, y) {
+export function set(out, x, y) {
     out[0] = x;
     out[1] = y;
     return out;
@@ -100,7 +99,7 @@ vec2.set = function(out, x, y) {
  * @param {vec2} b the second operand
  * @returns {vec2} out
  */
-vec2.add = function(out, a, b) {
+export function add(out, a, b) {
     out[0] = a[0] + b[0];
     out[1] = a[1] + b[1];
     return out;
@@ -114,17 +113,17 @@ vec2.add = function(out, a, b) {
  * @param {vec2} b the second operand
  * @returns {vec2} out
  */
-vec2.subtract = function(out, a, b) {
+export function subtract(out, a, b) {
     out[0] = a[0] - b[0];
     out[1] = a[1] - b[1];
     return out;
 };
 
 /**
- * Alias for {@link vec2.subtract}
+ * Alias for {@link subtract}
  * @function
  */
-vec2.sub = vec2.subtract;
+export const sub = subtract;
 
 /**
  * Multiplies two vec2's
@@ -134,17 +133,17 @@ vec2.sub = vec2.subtract;
  * @param {vec2} b the second operand
  * @returns {vec2} out
  */
-vec2.multiply = function(out, a, b) {
+export function multiply(out, a, b) {
     out[0] = a[0] * b[0];
     out[1] = a[1] * b[1];
     return out;
 };
 
 /**
- * Alias for {@link vec2.multiply}
+ * Alias for {@link multiply}
  * @function
  */
-vec2.mul = vec2.multiply;
+export const mul = multiply;
 
 /**
  * Divides two vec2's
@@ -154,17 +153,17 @@ vec2.mul = vec2.multiply;
  * @param {vec2} b the second operand
  * @returns {vec2} out
  */
-vec2.divide = function(out, a, b) {
+export function divide(out, a, b) {
     out[0] = a[0] / b[0];
     out[1] = a[1] / b[1];
     return out;
 };
 
 /**
- * Alias for {@link vec2.divide}
+ * Alias for {@link divide}
  * @function
  */
-vec2.div = vec2.divide;
+export const div = divide;
 
 /**
  * Math.ceil the components of a vec2
@@ -173,7 +172,7 @@ vec2.div = vec2.divide;
  * @param {vec2} a vector to ceil
  * @returns {vec2} out
  */
-vec2.ceil = function (out, a) {
+export function ceil(out, a) {
     out[0] = Math.ceil(a[0]);
     out[1] = Math.ceil(a[1]);
     return out;
@@ -186,7 +185,7 @@ vec2.ceil = function (out, a) {
  * @param {vec2} a vector to floor
  * @returns {vec2} out
  */
-vec2.floor = function (out, a) {
+export function floor(out, a) {
     out[0] = Math.floor(a[0]);
     out[1] = Math.floor(a[1]);
     return out;
@@ -200,7 +199,7 @@ vec2.floor = function (out, a) {
  * @param {vec2} b the second operand
  * @returns {vec2} out
  */
-vec2.min = function(out, a, b) {
+export function min(out, a, b) {
     out[0] = Math.min(a[0], b[0]);
     out[1] = Math.min(a[1], b[1]);
     return out;
@@ -214,7 +213,7 @@ vec2.min = function(out, a, b) {
  * @param {vec2} b the second operand
  * @returns {vec2} out
  */
-vec2.max = function(out, a, b) {
+export function max(out, a, b) {
     out[0] = Math.max(a[0], b[0]);
     out[1] = Math.max(a[1], b[1]);
     return out;
@@ -227,7 +226,7 @@ vec2.max = function(out, a, b) {
  * @param {vec2} a vector to round
  * @returns {vec2} out
  */
-vec2.round = function (out, a) {
+export function round(out, a) {
     out[0] = Math.round(a[0]);
     out[1] = Math.round(a[1]);
     return out;
@@ -241,7 +240,7 @@ vec2.round = function (out, a) {
  * @param {Number} b amount to scale the vector by
  * @returns {vec2} out
  */
-vec2.scale = function(out, a, b) {
+export function scale(out, a, b) {
     out[0] = a[0] * b;
     out[1] = a[1] * b;
     return out;
@@ -256,7 +255,7 @@ vec2.scale = function(out, a, b) {
  * @param {Number} scale the amount to scale b by before adding
  * @returns {vec2} out
  */
-vec2.scaleAndAdd = function(out, a, b, scale) {
+export function scaleAndAdd(out, a, b, scale) {
     out[0] = a[0] + (b[0] * scale);
     out[1] = a[1] + (b[1] * scale);
     return out;
@@ -269,17 +268,17 @@ vec2.scaleAndAdd = function(out, a, b, scale) {
  * @param {vec2} b the second operand
  * @returns {Number} distance between a and b
  */
-vec2.distance = function(a, b) {
+export function distance(a, b) {
     var x = b[0] - a[0],
         y = b[1] - a[1];
     return Math.sqrt(x*x + y*y);
 };
 
 /**
- * Alias for {@link vec2.distance}
+ * Alias for {@link distance}
  * @function
  */
-vec2.dist = vec2.distance;
+export const dist = distance;
 
 /**
  * Calculates the squared euclidian distance between two vec2's
@@ -288,17 +287,17 @@ vec2.dist = vec2.distance;
  * @param {vec2} b the second operand
  * @returns {Number} squared distance between a and b
  */
-vec2.squaredDistance = function(a, b) {
+export function squaredDistance(a, b) {
     var x = b[0] - a[0],
         y = b[1] - a[1];
     return x*x + y*y;
 };
 
 /**
- * Alias for {@link vec2.squaredDistance}
+ * Alias for {@link squaredDistance}
  * @function
  */
-vec2.sqrDist = vec2.squaredDistance;
+export const sqrDist = squaredDistance;
 
 /**
  * Calculates the length of a vec2
@@ -306,17 +305,17 @@ vec2.sqrDist = vec2.squaredDistance;
  * @param {vec2} a vector to calculate length of
  * @returns {Number} length of a
  */
-vec2.length = function (a) {
+export function length(a) {
     var x = a[0],
         y = a[1];
     return Math.sqrt(x*x + y*y);
 };
 
 /**
- * Alias for {@link vec2.length}
+ * Alias for {@link length}
  * @function
  */
-vec2.len = vec2.length;
+export const len = length;
 
 /**
  * Calculates the squared length of a vec2
@@ -324,17 +323,17 @@ vec2.len = vec2.length;
  * @param {vec2} a vector to calculate squared length of
  * @returns {Number} squared length of a
  */
-vec2.squaredLength = function (a) {
+export function squaredLength(a) {
     var x = a[0],
         y = a[1];
     return x*x + y*y;
 };
 
 /**
- * Alias for {@link vec2.squaredLength}
+ * Alias for {@link squaredLength}
  * @function
  */
-vec2.sqrLen = vec2.squaredLength;
+export const sqrLen = squaredLength;
 
 /**
  * Negates the components of a vec2
@@ -343,7 +342,7 @@ vec2.sqrLen = vec2.squaredLength;
  * @param {vec2} a vector to negate
  * @returns {vec2} out
  */
-vec2.negate = function(out, a) {
+export function negate(out, a) {
     out[0] = -a[0];
     out[1] = -a[1];
     return out;
@@ -356,7 +355,7 @@ vec2.negate = function(out, a) {
  * @param {vec2} a vector to invert
  * @returns {vec2} out
  */
-vec2.inverse = function(out, a) {
+export function inverse(out, a) {
   out[0] = 1.0 / a[0];
   out[1] = 1.0 / a[1];
   return out;
@@ -369,7 +368,7 @@ vec2.inverse = function(out, a) {
  * @param {vec2} a vector to normalize
  * @returns {vec2} out
  */
-vec2.normalize = function(out, a) {
+export function normalize(out, a) {
     var x = a[0],
         y = a[1];
     var len = x*x + y*y;
@@ -389,7 +388,7 @@ vec2.normalize = function(out, a) {
  * @param {vec2} b the second operand
  * @returns {Number} dot product of a and b
  */
-vec2.dot = function (a, b) {
+export function dot(a, b) {
     return a[0] * b[0] + a[1] * b[1];
 };
 
@@ -402,7 +401,7 @@ vec2.dot = function (a, b) {
  * @param {vec2} b the second operand
  * @returns {vec3} out
  */
-vec2.cross = function(out, a, b) {
+export function cross(out, a, b) {
     var z = a[0] * b[1] - a[1] * b[0];
     out[0] = out[1] = 0;
     out[2] = z;
@@ -418,7 +417,7 @@ vec2.cross = function(out, a, b) {
  * @param {Number} t interpolation amount between the two inputs
  * @returns {vec2} out
  */
-vec2.lerp = function (out, a, b, t) {
+export function lerp(out, a, b, t) {
     var ax = a[0],
         ay = a[1];
     out[0] = ax + t * (b[0] - ax);
@@ -433,9 +432,9 @@ vec2.lerp = function (out, a, b, t) {
  * @param {Number} [scale] Length of the resulting vector. If ommitted, a unit vector will be returned
  * @returns {vec2} out
  */
-vec2.random = function (out, scale) {
+export function random(out, scale) {
     scale = scale || 1.0;
-    var r = glMatrix.RANDOM() * 2.0 * Math.PI;
+    var r = RANDOM() * 2.0 * Math.PI;
     out[0] = Math.cos(r) * scale;
     out[1] = Math.sin(r) * scale;
     return out;
@@ -449,7 +448,7 @@ vec2.random = function (out, scale) {
  * @param {mat2} m matrix to transform with
  * @returns {vec2} out
  */
-vec2.transformMat2 = function(out, a, m) {
+export function transformMat2(out, a, m) {
     var x = a[0],
         y = a[1];
     out[0] = m[0] * x + m[2] * y;
@@ -465,7 +464,7 @@ vec2.transformMat2 = function(out, a, m) {
  * @param {mat2d} m matrix to transform with
  * @returns {vec2} out
  */
-vec2.transformMat2d = function(out, a, m) {
+export function transformMat2d(out, a, m) {
     var x = a[0],
         y = a[1];
     out[0] = m[0] * x + m[2] * y + m[4];
@@ -482,7 +481,7 @@ vec2.transformMat2d = function(out, a, m) {
  * @param {mat3} m matrix to transform with
  * @returns {vec2} out
  */
-vec2.transformMat3 = function(out, a, m) {
+export function transformMat3(out, a, m) {
     var x = a[0],
         y = a[1];
     out[0] = m[0] * x + m[3] * y + m[6];
@@ -500,8 +499,8 @@ vec2.transformMat3 = function(out, a, m) {
  * @param {mat4} m matrix to transform with
  * @returns {vec2} out
  */
-vec2.transformMat4 = function(out, a, m) {
-    var x = a[0], 
+export function transformMat4(out, a, m) {
+    var x = a[0],
         y = a[1];
     out[0] = m[0] * x + m[4] * y + m[12];
     out[1] = m[1] * x + m[5] * y + m[13];
@@ -520,8 +519,8 @@ vec2.transformMat4 = function(out, a, m) {
  * @returns {Array} a
  * @function
  */
-vec2.forEach = (function() {
-    var vec = vec2.create();
+export const forEach = (function() {
+    var vec = create();
 
     return function(a, stride, offset, count, fn, arg) {
         var i, l;
@@ -532,7 +531,7 @@ vec2.forEach = (function() {
         if(!offset) {
             offset = 0;
         }
-        
+
         if(count) {
             l = Math.min((count * stride) + offset, a.length);
         } else {
@@ -544,7 +543,7 @@ vec2.forEach = (function() {
             fn(vec, vec, arg);
             a[i] = vec[0]; a[i+1] = vec[1];
         }
-        
+
         return a;
     };
 })();
@@ -555,7 +554,7 @@ vec2.forEach = (function() {
  * @param {vec2} a vector to represent as a string
  * @returns {String} string representation of the vector
  */
-vec2.str = function (a) {
+export function str(a) {
     return 'vec2(' + a[0] + ', ' + a[1] + ')';
 };
 
@@ -566,7 +565,7 @@ vec2.str = function (a) {
  * @param {vec2} b The second vector.
  * @returns {Boolean} True if the vectors are equal, false otherwise.
  */
-vec2.exactEquals = function (a, b) {
+export function exactEquals(a, b) {
     return a[0] === b[0] && a[1] === b[1];
 };
 
@@ -577,11 +576,9 @@ vec2.exactEquals = function (a, b) {
  * @param {vec2} b The second vector.
  * @returns {Boolean} True if the vectors are equal, false otherwise.
  */
-vec2.equals = function (a, b) {
+export function equals(a, b) {
     var a0 = a[0], a1 = a[1];
     var b0 = b[0], b1 = b[1];
-    return (Math.abs(a0 - b0) <= glMatrix.EPSILON*Math.max(1.0, Math.abs(a0), Math.abs(b0)) &&
-            Math.abs(a1 - b1) <= glMatrix.EPSILON*Math.max(1.0, Math.abs(a1), Math.abs(b1)));
+    return (Math.abs(a0 - b0) <= EPSILON*Math.max(1.0, Math.abs(a0), Math.abs(b0)) &&
+            Math.abs(a1 - b1) <= EPSILON*Math.max(1.0, Math.abs(a1), Math.abs(b1)));
 };
-
-module.exports = vec2;
